usbtunnel: fix SETUP-OUT-IN by implementing my own queue for control
in the SETUP-OUT-IN case, once the completion callback for the OUT is
called, the udc emits the IN on its own. this means that we might get
a SETUP while mirroring the transaction to the device, making a huge
mess. we can't change the udc's behaviour (all udcs behave this way),
so instead a queue was implemented to delay handling these SETUPs
until the current transcation is complete.

Signed-off-by: Tal Shorer <tal.shorer@gmail.com>